Add parsing of arinc653.aadl
This is needed if user uses a custom AADL file in the command line
For example if user set this flag: --no-stdlib then
ocarina_components.aadl will not be parsed. User in that case has to
provide a custom library of components, like the one used for the AIR
plaform. But when an extra aadl file is passed in the command line,
kazoo does not know if it references Interface or Deployment components
so it parses it together with both Interface and Deployment views.
The deployment view parser already parses arinc653.aadl, but by default
the interface view parser did not (that is normally not needed)