Fix of memory alignment issue.
The buffers used when Remote Interface is called with Encoding NATIVE,
(for functions Decode_NATIVE_* and Encode_NATIVE_*)
should have the same alignment as parameter type.
Instead of using static char array, the buffer is allocated as a
variable with parameter type, so the compiler will guarantee valid
memory alignment.